requirements.txt
annotated-types==0.7.0
anyio==4.6.0
dnspython==2.6.1
exceptiongroup==1.2.2
fastapi==0.115.0
idna==3.10
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pymongo==4.6.1
pytest==8.3.2
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.38.6
tomli==2.0.1
typing_extensions==4.12.2
uvicorn
passlib
python-jose
python-multipart
httpx==0.27.2
